02-python-requests/requirements.txt
# Dependencies for FortiManager API with requests
requests>=2.28.0
python-dotenv>=1.0.0
urllib3>=2.0.0
